+++ src/ExtendedPromiseInterface.php
@@ -5,4 +5,5 @@ namespace React\Promise;
 interface ExtendedPromiseInterface extends PromiseInterface
 {
     public function done(callable $onFulfilled = null, callable $onRejected = null, callable $onProgress = null);
+    public function progress(callable $onProgress);
 }